requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
newspaper3k
lxml
lxml_html_clean
python-multipart
